+++ aten/src/ATen/Dispatch.h
@@ -5,7 +5,7 @@
 #include <c10/macros/Macros.h>
 #include <c10/util/Exception.h>
 #include <c10/util/Half.h>
-#include <c10/util/complex_type.h>
+#include <c10/util/complex.h>
 
 #define AT_PRIVATE_CASE_TYPE(enum_type, type, ...) \
   case enum_type: {                                \
