+++ aten/src/ATen/cpu/vml.h
@@ -4,7 +4,7 @@
 #include <ATen/Parallel.h>
 #include <ATen/cpu/vec256/functional.h>
 #include <ATen/cpu/vec256/vec256.h>
-#include <c10/util/complex_type.h>
+#include <c10/util/complex.h>
 
 // This header implements various unary operations using a MKL VML style
 // interface.
