+++ aten/src/ATen/native/cpu/zmath.h
@@ -1,7 +1,7 @@
 #pragma once
 
 // Complex number math operations that act as no-ops for other dtypes.
-#include <c10/util/complex_type.h>
+#include <c10/util/complex.h>
 #include <c10/util/math_compat.h>
 #include<ATen/NumericUtils.h>
