+++ c10/core/ScalarType.h
@@ -1,12 +1,12 @@
 #pragma once
 
 #include <c10/util/ArrayRef.h>
-#include <c10/util/complex_type.h>
+#include <c10/util/complex.h>
 #include <c10/util/Half.h>
 #include <c10/util/BFloat16.h>
 #include <c10/util/Optional.h>
 #include <c10/util/typeid.h>
-#include <c10/util/complex_type.h>
+#include <c10/util/complex.h>
 
 #include <complex>
 #include <cstdint>
